% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecld-pdf-method.R
\name{ecld.pdf}
\alias{ecld.pdf}
\title{Calculate the PDF of an ecld object}
\usage{
ecld.pdf(object, x)
}
\arguments{
\item{object}{an object of ecd class}

\item{x}{numeric vector of \eqn{x} dimension}
}
\value{
numeric vector of the PDF
}
\description{
Calculate the PDF of an ecld object
}
\examples{
ld <- ecld(lambda=3)
x <- seq(-10, 10, by=1)
ecld.pdf(ld,x)
}
\author{
Stephen H-T. Lihn
}
\keyword{distribution}
\keyword{pdf}
