% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecld-constructor.R
\name{ecld}
\alias{ecld}
\title{Constructor of ecld class}
\usage{
ecld(lambda = 3, sigma = 1, mu = 0, verbose = FALSE)
}
\arguments{
\item{lambda}{numeric, the lambda parameter. Must be positive. Default: 3.}

\item{sigma}{numeric, the scale parameter. Must be positive. Default: 1.}

\item{mu}{numeric, the location parameter. Default: 0.}

\item{verbose}{logical, display timing information, for debugging purpose, default is \code{FALSE}.}
}
\value{
an object of ecld class
}
\description{
Construct an \code{\link{ecld-class}} by providing the required parameters.
The default is the standard symmetric cusp distribution (lambda=3).
}
\examples{
ld <- ecld()
ld <- ecld(2, 0.01)
}
\author{
Stephen H-T. Lihn
}
\keyword{constructor}
