% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecld-sd-method.R
\name{ecld.sd}
\alias{ecld.sd}
\alias{ecld.var}
\alias{ecld.mean}
\alias{ecld.skewness}
\alias{ecld.kurtosis}
\alias{ecld.kurt}
\title{Compute statistics analytically for an ecld object}
\usage{
ecld.sd(object)

ecld.var(object)

ecld.mean(object)

ecld.skewness(object)

ecld.kurtosis(object)

ecld.kurt(object)
}
\arguments{
\item{object}{an object of ecld class}
}
\value{
numeric
}
\description{
Compute statistics for mean, var, skewness, kurtosis.
}
\examples{
ld <- ecld(3)
ecld.sd(ld)
ecld.var(ld)
ecld.mean(ld)
ecld.skewness(ld)
ecld.kurt(ld)

}
\author{
Stephen H-T. Lihn
}
\keyword{statistics}
