% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldhmm-conditional_prob.R
\name{ldhmm.conditional_prob}
\alias{ldhmm.conditional_prob}
\title{Computing the conditional probabilities}
\usage{
ldhmm.conditional_prob(object, x, xc)
}
\arguments{
\item{object}{an ldhmm object}

\item{x}{numeric, the observations.}

\item{xc}{numeric, the conditional observations.}
}
\value{
matrix of probabilities, size of xc times size of x.
}
\description{
This utility computes the conditional probabilities that observation at time t 
equals xc, given all observations other than that at time t being the same.
}
\author{
Stephen H. Lihn
}
\keyword{pdf}

