% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldhmm-decoding.R
\name{ldhmm.decoding}
\alias{ldhmm.decoding}
\title{Computing the minus log-likelihood (MLLK)}
\usage{
ldhmm.decoding(object, x)
}
\arguments{
\item{object}{an ldhmm object}

\item{x}{numeric, the observations.}
}
\value{
an ldhmm object containing results of decoding
}
\description{
This utility computes the state probabilities, uses local and global decoding to calculate the states.
The results are saved to the returned \code{ldhmm} object.
}
\author{
Stephen H. Lihn
}
\keyword{mllk}
