% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldhmm-forecast_prob.R
\name{ldhmm.forecast_prob}
\alias{ldhmm.forecast_prob}
\title{Computing the forecast probability distribution}
\usage{
ldhmm.forecast_prob(object, x, xf, h = 1)
}
\arguments{
\item{object}{an ldhmm object}

\item{x}{numeric, the observations.}

\item{xf}{numeric, the future observations to be forecasted.}

\item{h}{integer, time steps to forecast.}
}
\value{
matrix of probabilities, size of h times size of xf.
}
\description{
This utility computes the forecast probability distribution (Zucchini, 5.3)
}
\author{
Stephen H. Lihn
}
\keyword{forecast}

