% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldhmm-forecast_state.R
\name{ldhmm.forecast_state}
\alias{ldhmm.forecast_state}
\title{Computing the state forecast}
\usage{
ldhmm.forecast_state(object, x, h = 1)
}
\arguments{
\item{object}{an ldhmm object}

\item{x}{numeric, the observations.}

\item{h}{integer, time steps to forecast.}
}
\value{
matrix of probabilities per state (even if h=1), number of states times size of h
}
\description{
This utility computes the state forecast, given the sequence of observations in the past.
}
\author{
Stephen H. Lihn
}
\keyword{forecast}
