% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldhmm-forecast_volatility.R
\name{ldhmm.forecast_volatility}
\alias{ldhmm.forecast_volatility}
\title{Computing the volatility forecast for next one period}
\usage{
ldhmm.forecast_volatility(object, x, xf, ma.order = 0, days.pa = 252)
}
\arguments{
\item{object}{an ldhmm object}

\item{x}{numeric, the observations.}

\item{xf}{numeric, the future observations to be forecasted.}

\item{ma.order}{a positive integer or zero, specifying order of moving average. Default is zero.}

\item{days.pa}{a positive integer specifying trading days per year, default is 252.}
}
\value{
matrix of future observations and volatilities, size of 2 times length of \code{xf}.
}
\description{
This utility computes the volatility forecast based on the given future observations for next one period.
}
\author{
Stephen H. Lihn
}
\keyword{forecast}
