% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldhmm-ld_stats.R
\name{ldhmm.ld_stats}
\alias{ldhmm.ld_stats}
\title{Computes the theoretical statistics per state}
\usage{
ldhmm.ld_stats(object, annualize = FALSE, days.pa = 252)
}
\arguments{
\item{object}{an ldhmm object}

\item{annualize}{logical, to annaulize the sd and mean to V (xsqrt(days.pa)x100) and R (xdays.pa). 
Default is \code{FALSE}.}

\item{days.pa}{a positive integer, specifying number of days per year, default is 252.}
}
\value{
a matrix of statistics for each state, size of states times 4
}
\description{
This utility computes the statistics (mean, sd, kurtosis) based on the lambda distribution.
This is used to compare to the statistics from observations for each state.
\code{alloc} is a short-hand for Merton's optimal allocation, \code{mean/sd^2}.
}
\author{
Stephen H. Lihn
}
\keyword{pdf}
