% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldhmm-mllk.R
\name{ldhmm.mllk}
\alias{ldhmm.mllk}
\title{Computing the minus log-likelihood (MLLK)}
\usage{
ldhmm.mllk(par.vector, x, ldhmm, mllk.print.level = 0)
}
\arguments{
\item{par.vector}{numeric, the working parameter vector}

\item{x}{numeric, the observations.}

\item{ldhmm}{an input ldhmm object to provide static reference, 
such as m, param.nbr, stationary.}

\item{mllk.print.level}{numeric, this argument determines the level of printing 
which is done during the minimization process. 
The default value of 0 means that no printing occurs, 
a value of 1 or greater means some tracing information is printed.}
}
\value{
an ldhmm object containing results of MLE optimization
}
\description{
This utility computes the MLLK. It is typically invoked by the MLE optimizer. (Zucchini, 3.2)
}
\author{
Stephen H. Lihn
}
\keyword{mllk}

