% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldhmm-oxford_man_index_list.R
\name{ldhmm.oxford_man_index_list}
\alias{ldhmm.oxford_man_index_list}
\title{Get the index list from Oxford-Man}
\usage{
ldhmm.oxford_man_index_list()
}
\value{
character, list of indices
}
\description{
This utility shows the index list within the Oxford-Man data set.
}
\examples{
\dontrun{
    ldhmm.oxford_man_index_list()
}
}
\references{
Oxford-Man Institute of Quantitative Finance. Realized Library: http://realized.oxford-man.ox.ac.uk
}
\author{
Stephen H. Lihn
}
\keyword{oxford}
