% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldhmm-read-csv-by-symbol-method.R
\name{ldhmm.read_csv_by_symbol}
\alias{ldhmm.read_csv_by_symbol}
\title{Read csv file of sample data}
\usage{
ldhmm.read_csv_by_symbol(symbol = "dji", extdata_dir = NULL)
}
\arguments{
\item{symbol}{Character for the symbol of the time series. Default: dji}

\item{extdata_dir}{optionally specify user's own extdata folder}
}
\value{
The data.frame object
}
\description{
This is a helper utility to read sample csv file into data frame.
The main use for external users is to read the option data
since it has a different format than other price timeseries data.
}
\examples{
dji <- ldhmm.read_csv_by_symbol("dji")
spx <- ldhmm.read_csv_by_symbol("spx")

}
\author{
Stephen H-T. Lihn
}
\keyword{data}
