% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldhmm-read_sample_object.R
\name{ldhmm.read_sample_object}
\alias{ldhmm.read_sample_object}
\title{Read sample ldhmm object}
\usage{
ldhmm.read_sample_object(symbol = "spx-daily-m10", extdata_dir = NULL)
}
\arguments{
\item{symbol}{Character for the symbol of the time series. Default is \code{spx-daily-m10}}

\item{extdata_dir}{optionally specify user's own extdata folder}
}
\value{
The ldhmm object
}
\description{
This utility is used to read sample ldhmm object so that the user doesn't need to go through
lengthy optimization process to obtain a trained HMM for advanced features.
}
\examples{
hs <- ldhmm.read_sample_object() # SPX daily 10-state HMM

}
\author{
Stephen H-T. Lihn
}
\keyword{data}
