% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldhmm-simulate_abs_acf.R
\name{ldhmm.simulate_abs_acf}
\alias{ldhmm.simulate_abs_acf}
\title{Simulating auto-correlation (ACF)}
\usage{
ldhmm.simulate_abs_acf(object, n = 10000, lag.max = 5, debug = FALSE)
}
\arguments{
\item{object}{an ldhmm object that can supply m, param.nbr and stationary.}

\item{n}{a positive integer specifying number of observations to simulate.}

\item{lag.max}{a positive integer, specifying number of lags to be computed.}

\item{debug}{logical, specifying to print progress message or not. Default is \code{FALSE}.}
}
\value{
a vector of ACF
}
\description{
This utility simulates the auto-correlation. The first few lag of ACF should match 
the ACF from the market data fairly well. This is a major validation of a successful HMM.
Be aware this is a CPU intensive calculation. It uses the multi-core functionality.
}
\author{
Stephen H. Lihn
}
\keyword{acf}
