% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldhmm-simulate_state_transition.R
\name{ldhmm.simulate_state_transition}
\alias{ldhmm.simulate_state_transition}
\title{Simulating state transition}
\usage{
ldhmm.simulate_state_transition(object, init = NULL)
}
\arguments{
\item{object}{an ldhmm object that can supply m, param.nbr and stationary.}

\item{init}{a positive integer specifying number of observations to simulate initially.
The default is NULL, indicating that the simulation should use the (local) states and observations
from within the object, and simulate the next set of random states and observations according to gamma.
When init is an integer, the utility will generate random states and observations according to delta.}
}
\value{
an ldhmm object containing the simulated states and observations.
        The observations are stored in the \code{observations} slot.
        The states are stored in the \code{states.local} slot.
}
\description{
This utility allows to simulate the states and obervations over time. 
Be aware this is a CPU intensive calculation. It uses the multi-core functionality.
}
\author{
Stephen H. Lihn
}
\keyword{simulation}
