% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldhmm-state_ld.R
\name{ldhmm.state_ld}
\alias{ldhmm.state_ld}
\title{Constructing the ecld objects per state}
\usage{
ldhmm.state_ld(object, state = NULL)
}
\arguments{
\item{object}{an ldhmm object}

\item{state}{numeric, the states.}
}
\value{
a list of ecld objects
}
\description{
This utility constructs the ecld objects per state and return them in a list of easy query.
}
\author{
Stephen H. Lihn
}
\keyword{pdf}

