% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldhmm-state_pdf.R
\name{ldhmm.state_pdf}
\alias{ldhmm.state_pdf}
\title{Computing the PDF per state given the observations}
\usage{
ldhmm.state_pdf(object, state, x)
}
\arguments{
\item{object}{an ldhmm object}

\item{state}{numeric, the states.}

\item{x}{numeric, the observations.}
}
\value{
a vector or matrix of PDF. The dimension of matrix is state times x
}
\description{
Computing the PDF per state given the observations. Only one of state or x 
can be a vector per call.
}
\author{
Stephen H. Lihn
}
\keyword{pdf}
