% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldhmm-ts_abs_acf.R
\name{ldhmm.ts_abs_acf}
\alias{ldhmm.ts_abs_acf}
\title{Computing ACF of the absolute value of a time series}
\usage{
ldhmm.ts_abs_acf(x, drop = 0, lag.max = 100)
}
\arguments{
\item{x}{numeric, the observations.}

\item{drop}{a positive integer, specifying number of outliers to be dropped.}

\item{lag.max}{a positive integer, specifying number of lags to be computed.}
}
\value{
a vector of ACF
}
\description{
This utility computes the ACF of the absolute value of a time series as a proxy 
of the auto-correlation of the volatility. It allows to drop the largest N outliers
so that they would not skew the ACF calculation.
}
\author{
Stephen H. Lihn
}
\keyword{acf}
