% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldhmm-viterbi.R
\name{ldhmm.viterbi}
\alias{ldhmm.viterbi}
\title{Computing the global decoding by the Viterbi algorithm}
\usage{
ldhmm.viterbi(object, x)
}
\arguments{
\item{object}{an ldhmm object}

\item{x}{numeric, the observations.}
}
\value{
a vector of states
}
\description{
This utility computes the global decoding by the Viterbi algorithm.
}
\author{
Stephen H. Lihn
}
\keyword{viterbi}
