% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldhmm-w2n.R
\name{ldhmm.w2n}
\alias{ldhmm.w2n}
\title{Transforming working parameter array to natural parameters}
\usage{
ldhmm.w2n(object, par.vector, mu.scale = 1)
}
\arguments{
\item{object}{an ldhmm object that can supply m, param.nbr and stationary.}

\item{par.vector}{numeric, linear working parameter array. See \code{ldhmm.n2w}.}

\item{mu.scale}{numeric, it should mirror what is provided to \code{ldhmm.n2w}.
Default is 1.}
}
\value{
an ldhmm object
}
\description{
This utility transforms the working parameter array back to
the vectors and matrix of the contrained parameters. (Zucchini, 3.3.1)
}
\author{
Stephen H. Lihn
}
\keyword{parameter}
