% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldhmm-ld_stats.R
\name{ldhmm.ld_stats}
\alias{ldhmm.ld_stats}
\title{Computes the theoretical statistics per state}
\usage{
ldhmm.ld_stats(object)
}
\arguments{
\item{object}{an ldhmm object}
}
\value{
a matrix of statistics for each state, size of states times 3
}
\description{
This utility computes the statistics (mean, sd, kurtosis) based on the lambda distribution.
This is used to compare to the statistics from observations for each state.
}
\author{
Stephen H. Lihn
}
\keyword{pdf}

