% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldhmm-oxford_man_plot_obs.R
\name{ldhmm.oxford_man_plot_obs}
\alias{ldhmm.oxford_man_plot_obs}
\title{Plotting Oxford-Man realized volatility overlaid with HMM expected volatility}
\usage{
ldhmm.oxford_man_plot_obs(object, days.pa = 252, start.date = NULL,
  end.date = NULL, index.r = "SPX2.r", index.rv = "SPX2.rv",
  index.px = "SPX2.closeprice", index.px.scale = 15,
  index.px.origin = NULL, index.vol.ma.order = 5, index.symbol = NULL,
  vix.adj.ratio = NULL, insert.plot = TRUE, insert.viewport = NULL)
}
\arguments{
\item{object}{an ldhmm object with a stationary solution. If this is set to \code{NULL},
an internal 10-state HMM object will be used.}

\item{days.pa}{a positive integer specifying trading days per year, default is 252.}

\item{start.date}{Date or character of ISO format (YYYY-MM-DD), 
specifying the start date of the plot, default is \code{NULL}.}

\item{end.date}{Date or character of ISO format (YYYY-MM-DD), 
specifying the end date of the plot, default is \code{NULL}.}

\item{index.r}{character, specifying index return column, default is \code{SPX2.r}.}

\item{index.rv}{character, specifying index realized variance column, default is \code{SPX2.rv}.}

\item{index.px}{character, specifying index closing price column, default to \code{"SPX2.closeprice"}.
Set this to \code{NULL} if you don't wish to see the price line.}

\item{index.px.scale}{numeric, specifying the scaling factor when plotting price trend, default is 15.
The closing price is converted to cumulative return by the price of the first date.
Then plot from the mid-point of volatility axis with this scale.}

\item{index.px.origin}{numeric, specifying the starting value of the index price line,
the default is \code{NULL}, which will start the index price line from the middle of y-axis.}

\item{index.vol.ma.order}{a positive integer specifying the simple moving average of the realized volatility, default is 5.
This is needed because the realized volatility is very noisy at the daily level.}

\item{index.symbol}{character, used as a shortcut for \code{index.r, index.rv, index.px}. Once the root symbol is specified,
the three time series can be derived by appending suffix \code{.r, .rv, .closeprice}. 
This mechanism will supercede those three components 
if and only if a value is detected. Default is \code{NULL}.}

\item{vix.adj.ratio}{numeric, if specified, VIX index is adjusted and plotted, default is \code{NULL}.
The long-term ratio between VIX and 10-state HMM is 0.79.
The VIX data is cached when the Oxford data is downloaded.}

\item{insert.plot}{logical, if true, also plot the volatility-return as insert in upper-right corner, default is \code{TRUE}.}

\item{insert.viewport}{optional viewport for the insert, default is \code{NULL}, 
which is internally set to \code{grid::viewport(.8, .75, .3, .3)}.}
}
\description{
This utility plots the Oxford-Man realized volatility (from SPX2.rv) and overlays with the HMM expected volatility
with the observations set up SPX2.r. This graph is to show that the HMM is capable of reproducing the realized volatility.
Optionally the insert shows the relation between the return and volatility indicated by each state. This plot is also called
"volatility yield curve".
}
\examples{
\dontrun{
    ldhmm.oxford_man_plot_obs(h)
}
}
\author{
Stephen H. Lihn
}
\keyword{oxford}
